♻️ extract mockClock
