+++ packages/core/test/index.ts
@@ -4,6 +4,7 @@ export * from './capturedExceptions'
 export * from './collectAsyncCalls'
 export * from './createNewEvent'
 export * from './location.specHelper'
+export * from './mockClock'
 export * from './requests.specHelper'
 export * from './stubReportApis'
 export * from './stubZoneJs'
