+++ packages/core/test/mockClock.ts
@@ -1,0 +1,20 @@
+import { resetNavigationStart } from '../src/tools/timeUtils'
+
+export type Clock = ReturnType<typeof mockClock>
+
+export function mockClock(date?: Date) {
+  jasmine.clock().install()
+  jasmine.clock().mockDate(date)
+  const start = Date.now()
+  spyOn(performance, 'now').and.callFake(() => Date.now() - start)
+  spyOnProperty(performance.timing, 'navigationStart', 'get').and.callFake(() => start)
+  resetNavigationStart()
+  return {
+    tick: (ms: number) => jasmine.clock().tick(ms),
+    setDate: (date: Date) => jasmine.clock().mockDate(date),
+    cleanup: () => {
+      jasmine.clock().uninstall()
+      resetNavigationStart()
+    },
+  }
+}
