+++ packages/core/test/specHelper.ts
@@ -1,26 +1,6 @@
 import { instrumentMethod } from '../src/tools/instrumentMethod'
-import { resetNavigationStart } from '../src/tools/timeUtils'
 import { noop } from '../src/tools/utils'
 import type { BrowserWindowWithEventBridge } from '../src/transport'
-
-export type Clock = ReturnType<typeof mockClock>
-
-export function mockClock(date?: Date) {
-  jasmine.clock().install()
-  jasmine.clock().mockDate(date)
-  const start = Date.now()
-  spyOn(performance, 'now').and.callFake(() => Date.now() - start)
-  spyOnProperty(performance.timing, 'navigationStart', 'get').and.callFake(() => start)
-  resetNavigationStart()
-  return {
-    tick: (ms: number) => jasmine.clock().tick(ms),
-    setDate: (date: Date) => jasmine.clock().mockDate(date),
-    cleanup: () => {
-      jasmine.clock().uninstall()
-      resetNavigationStart()
-    },
-  }
-}
 
 export function setPageVisibility(visibility: 'visible' | 'hidden') {
   Object.defineProperty(document, 'visibilityState', {
